Derive Formulae From Augmented DAG

Uses the form attribute of edges in an augmented DAG to construct formulae for the relationship between exposure and outcome.

Usage

derive_formula(x, exposure, outcome, data = NULL, ...)

Arguments

x

An object of class dagitty.

exposure

Atomic character, indicating the exposure node in x.

outcome

Atomic character, indicating the outcome node in x.

data

Optional, a data.frame to be used as environment for the formulae. Default: NULL

...

Additional arguments, passed to adjustmentSets

Value

A list of objects with class formula.

Details

The form attribute of an augmented DAG of class dagitty should contain information about the functional form for the relationships specified in the DAG. For this function, the form attribute must be an additive function as also accepted by formula.

See also

Examples

x <- dagitty::dagitty('dag {
C
O
X
O <- X [form="X^2"]
C -> X
Y -> O
}')
derive_formula(x, outcome = "O", exposure = "X")
#> $`1`
#> O ~ X + X^2
#> NULL
#>