Create Vignette with webexercises Support

This function wraps rmarkdown::html_document to configure compilation to embed the default webexercises CSS and JavaScript files in the resulting HTML.

Usage

webex_vignette(...)

Arguments

...

Additional function arguments to pass to html_document.

Value

R Markdown output format to pass to 'render'.

Details

Call this function as the output_format argument for the render function when compiling HTML documents from RMarkdown source.

See also

Examples

# copy the webexercises 'R Markdown' template to a temporary file
if (FALSE) { # \dontrun{
my_rmd <- tempfile(fileext = ".Rmd")
rmarkdown::draft(my_rmd, "webexercises", "webexercises")

# compile it
rmarkdown::render(my_rmd, webexercises::webex_vignette())

# view the result
browseURL(sub("\\.Rmd$", ".html", my_rmd))
} # }